+++ src/ts/Conviva.d.ts
@@ -39,28 +39,47 @@ declare namespace Conviva {
 
   class Client {
     constructor(settings: ClientSettings, systemFactory: SystemFactory);
-    adEnd(sessionKey: number): void;
-    adStart(sessionKey: number, adStream: Client.AdStream, adPlayer: Client.AdPlayer,
-            adPosition: Client.AdPosition): void;
-    attachPlayer(sessionKey: number, playerStateManager: PlayerStateManager): void;
-    cleanupSession(sessionKey: number): void;
-    contentPreload(sessionKey: number): void;
-    contentStart(sessionKey: number): void;
-    createSession(contentMetadata: ContentMetadata | null): number;
-    detachPlayer(sessionKey: number): void;
-    getPlayerStateManager(): PlayerStateManager;
-    release(): void;
-    releasePlayerStateManager(playerStateManager: PlayerStateManager): void;
-    reportError(sessionKey: number, errorMessage: string, errorSeverity: Client.ErrorSeverity): void;
-    sendCustomEvent(sessionKey: number, eventName: string, eventAttributes: {}): void;
-    updateContentMetadata(sessionKey: number, contentMetadata: ContentMetadata): void;
+
+    public adEnd(sessionKey: number): void;
+
+    public adStart(
+      sessionKey: number,
+      adStream: Client.AdStream,
+      adPlayer: Client.AdPlayer,
+      adPosition: Client.AdPosition,
+    ): void;
+
+    public attachPlayer(sessionKey: number, playerStateManager: PlayerStateManager): void;
+
+    public cleanupSession(sessionKey: number): void;
+
+    public contentPreload(sessionKey: number): void;
+
+    public contentStart(sessionKey: number): void;
+
+    public createSession(contentMetadata: ContentMetadata | null): number;
+
+    public detachPlayer(sessionKey: number): void;
+
+    public getPlayerStateManager(): PlayerStateManager;
+
+    public release(): void;
+
+    public releasePlayerStateManager(playerStateManager: PlayerStateManager): void;
+
+    public reportError(sessionKey: number, errorMessage: string, errorSeverity: Client.ErrorSeverity): void;
+
+    public sendCustomEvent(sessionKey: number, eventName: string, eventAttributes: {}): void;
+
+    public updateContentMetadata(sessionKey: number, contentMetadata: ContentMetadata): void;
   }
 
   class ClientSettings {
+    public customerKey: string;
+    public gatewayUrl: string;
+    public heartbeatInterval: number;
+
     constructor(customerKey: string);
-    customerKey: string;
-    gatewayUrl: string;
-    heartbeatInterval: number;
   }
 
   namespace ContentMetadata {
@@ -72,48 +91,63 @@ declare namespace Conviva {
   }
 
   class ContentMetadata {
-    applicationName: string;
-    assetName: string;
-    custom: {};
-    defaultResource: string;
-    duration: number;
-    encodedFrameRate: number;
-    streamType: ContentMetadata.StreamType;
-    streamUrl: string;
-    viewerId: string;
+    public applicationName: string;
+    public assetName: string;
+    public custom: {};
+    public defaultResource: string;
+    public duration: number;
+    public encodedFrameRate: number;
+    public streamType: ContentMetadata.StreamType;
+    public streamUrl: string;
+    public viewerId: string;
   }
 
-  interface HttpRequestCancelFunction {
-    (): void;
-  }
+  type HttpRequestCancelFunction = () => void;
 
-  interface HttpRequestCallback {
-    (succeeded: boolean, data: string): void;
-  }
+  type HttpRequestCallback = (succeeded: boolean, data: string) => void;
 
   interface HttpInterface {
-    makeRequest(httpMethod: 'GET' | 'POST', url: string, data: string | null, contentType: string | null,
-                timeoutMs: number, callback: HttpRequestCallback | null): HttpRequestCancelFunction;
+    makeRequest(
+      httpMethod: 'GET' | 'POST',
+      url: string,
+      data: string | null,
+      contentType: string | null,
+      timeoutMs: number,
+      callback: HttpRequestCallback | null,
+    ): HttpRequestCancelFunction;
+
     release(): void;
   }
 
   interface LoggingInterface {
     consoleLog(message: string, logLevel: SystemSettings.LogLevel): void;
+
     release(): void;
   }
 
   interface MetadataInterface {
     getBrowserName(): string | null;
+
     getBrowserVersion(): string | null;
+
     getDeviceBrand(): string | null;
+
     getDeviceManufacturer(): string | null;
+
     getDeviceModel(): string | null;
+
     getDeviceType(): Client.DeviceType;
+
     getDeviceVersion(): string | null;
+
     getFrameworkName(): string | null;
+
     getFrameworkVersion(): string | null;
+
     getOperatingSystemName(): string | null;
+
     getOperatingSystemVersion(): string | null;
+
     release(): void;
   }
 
@@ -129,52 +163,78 @@ declare namespace Conviva {
   }
 
   class PlayerStateManager {
-    getBitrateKbps(): number;
-    getDuration(): number;
-    getEncodedFrameRate(): number;
-    getPlayerState(): PlayerStateManager.PlayerState;
-    getPlayerType(): string | null;
-    getPlayerVersion(): string | null;
-    getRenderedFrameRate(): number;
-    release(): void;
-    reset(): void;
-    sendError(errorMessage: string, errorSeverity: Client.ErrorSeverity): void;
-    setBitrateKbps(newBitrateKbps: number): void;
-    setDuration(duration: number): void;
-    setEncodedFrameRate(encodedFrameRate: number): void;
-    setPlayerSeekEnd(): void;
-    setPlayerSeekStart(seekToPos: number): void;
-    setPlayerState(newState: PlayerStateManager.PlayerState): void;
-    setPlayerType(playerType: string): void;
-    setPlayerVersion(playerVersion: string): void;
-    setRenderedFrameRate(renderedFrameRate: number): void;
-    setStreamUrl(streamUrl: string): void;
-    setUserSeekButtonDown(): void;
-    setUserSeekButtonUp(): void;
+    public getBitrateKbps(): number;
+
+    public getDuration(): number;
+
+    public getEncodedFrameRate(): number;
+
+    public getPlayerState(): PlayerStateManager.PlayerState;
+
+    public getPlayerType(): string | null;
+
+    public getPlayerVersion(): string | null;
+
+    public getRenderedFrameRate(): number;
+
+    public release(): void;
+
+    public reset(): void;
+
+    public sendError(errorMessage: string, errorSeverity: Client.ErrorSeverity): void;
+
+    public setBitrateKbps(newBitrateKbps: number): void;
+
+    public setDuration(duration: number): void;
+
+    public setEncodedFrameRate(encodedFrameRate: number): void;
+
+    public setPlayerSeekEnd(): void;
+
+    public setPlayerSeekStart(seekToPos: number): void;
+
+    public setPlayerState(newState: PlayerStateManager.PlayerState): void;
+
+    public setPlayerType(playerType: string): void;
+
+    public setPlayerVersion(playerVersion: string): void;
+
+    public setRenderedFrameRate(renderedFrameRate: number): void;
+
+    public setStreamUrl(streamUrl: string): void;
+
+    public setUserSeekButtonDown(): void;
+
+    public setUserSeekButtonUp(): void;
   }
 
-  interface StorageLoadDataCallback {
-    (succeeded: boolean, data: string| null): void;
-  }
+  type StorageLoadDataCallback = (succeeded: boolean, data: string | null) => void;
 
-  interface StorageSaveDataCallback {
-    (succeeded: boolean, data: string | null): void;
-  }
+  type StorageSaveDataCallback = (succeeded: boolean, data: string | null) => void;
 
   interface StorageInterface {
     loadData(storageSpace: string, storageKey: string, callback: StorageLoadDataCallback): void;
+
     saveData(storageSpace: string, storageKey: string, data: string, callback: StorageSaveDataCallback): void;
+
     release(): void;
   }
 
   class SystemFactory {
     constructor(systemInterface: SystemInterface, systemSettings: SystemSettings);
-    release(): void;
+
+    public release(): void;
   }
 
   class SystemInterface {
-    constructor(timeInterface: TimeInterface, timerInterface: TimerInterface, httpInterface: HttpInterface,
-        storageInterface: StorageInterface, metadataInterface: MetadataInterface, loggingInterface: LoggingInterface);
+    constructor(
+      timeInterface: TimeInterface,
+      timerInterface: TimerInterface,
+      httpInterface: HttpInterface,
+      storageInterface: StorageInterface,
+      metadataInterface: MetadataInterface,
+      loggingInterface: LoggingInterface,
+    );
   }
 
   namespace SystemSettings {
@@ -188,27 +248,25 @@ declare namespace Conviva {
   }
 
   class SystemSettings {
-    allowUncaughtExceptions: boolean;
-    httpTimeout: number;
-    logLevel: SystemSettings.LogLevel;
-    storageTimeout: number;
+    public allowUncaughtExceptions: boolean;
+    public httpTimeout: number;
+    public logLevel: SystemSettings.LogLevel;
+    public storageTimeout: number;
   }
 
   interface TimeInterface {
     getEpochTimeMs(): number;
+
     release(): void;
   }
 
-  interface TimerAction {
-    (): void;
-  }
+  type TimerAction = () => void;
 
-  interface TimerCancelFunction {
-    (): void;
-  }
+  type TimerCancelFunction = () => void;
 
   interface TimerInterface {
     createTimer(timerAction: TimerAction, intervalMs: number, actionName?: string | null): TimerCancelFunction;
+
     release(): void;
   }
 }