+++ webpack.dev.js
@@ -1,0 +1,16 @@
+const path = require('path');
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+  mode: 'development',
+  devtool: 'inline-source-map',
+  devServer: {
+    publicPath: '/dist/',
+    contentBase: path.resolve(__dirname, 'example')
+  },
+  optimization: {
+    nodeEnv: 'development',
+    minimize: false
+  }
+});